Create module bsh as BeanShell Assembly.
Within the new modularized project bsh is a packaging pom type module with no source used for assembling the separate modules back into a combined bsh jar again.
The assembly descriptor is the default `jar-with-dependencies` but with `useProjectArtifact` switched off because this module project does not provide its own artifacts.